+++ pkg/providers/property/provider.go
@@ -101,6 +101,7 @@ func (p *Subprovider) FrameworkDataSources() []func() datasource.DataSource {
 		NewDomainOwnershipSearchDomains,
 		NewHostnameActivationDataSource,
 		NewHostnameActivationsDataSource,
+		NewHostnameAuditHistoryDataSource,
 		NewHostnamesDiffDataSource,
 		NewIncludeDataSource,
 	}