+++ pkg/providers/property/testdata/TestDataPropertyHostnameAuditHistory/missing_hostname.tf
@@ -1,0 +1,6 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_property_hostname_audit_history" "test" {
+}